Fix incorrect loop condition in egg_hkdf_perform()
This does not cause a change in behavior (as evidenced by tests,
at least on linux when built with gcc) but is more correct code,
and less likely to be miscompiled.

The condition (i < 256) in the following loop is always false since i
is of type guchar, which is at most 255.

	guchar i;
	...
	for (i = 1; i < 256; ++i) { ... }

This patch changes i to a larger type gint.

Also in the loop we have:

	gcry_md_write (md2, &i, 1);

change it to use gcry_md_putc().